Change exclude-newer behavior to include indexes missing upload-time
Currently, when --exclude-newer is given, all indexes that are missing
it are treated entirely as unavailable. This makes it impossible to
use this flag in combination with many indexes.

The new behavior warns about the missing field, but treats the
distribution as available when the upload date it unknown.

This makes it possible to use this flag while still using packages
from indexes that don't support `upload-time`.